requirements.txt
torch
torchvision
timm # For Vision Transformers
albumentations
opencv-python
scikit-learn
matplotlib
numpy
Pillow
grad-cam # For interpretability